+++ doc/changelog.rst
@@ -1,3 +1,8 @@
+v0.6.1 - 04.11.2022
+====
+* New class: Base_animated_object
+* New initialization method to prepare list of sprites for further animation process
+
 v0.6.0 - 03.11.2022
 ====
 * New sprites for future animated units: spider tanks and fighters